This merge will introduce reproduction behavior for sev level 3 and 2 findings. Currently, there is no UI or method for invoking this feature because it will likely take a lot more time to get the updated UI working and will probably justify more UI changes. When a sev level 3 or 2 is discovered, the plugin automatically spins up a tab to open the page the event occured on and caches all responses from the page load. These responses are stored in an in-memory cache and are used to perform reproduction steps without changing the state of the application. It allows frontend fuzzing with multiple tabs in parallel that don't change the state of the app. Ideally, the use will click some button that will allow them to initate the reproductions.
starting new branch for reproduction steps

Signed-off-by: Jacob Ryan Heath <jake.heath@nccgroup.trust>

First step to autoreproduction steps finished. Base cases seem to
work, reflected XSS still needs to be worked on though.

Got the cache to properly store request responses and have removed all
cache headers so we don't accidently store 304 responses. Last thing
to hook up is the auto-prep on red and orange flags and the payload
swapping. This commit also fixed issues with gziping and transfer
encoding chunked. We don't store compressed or chunked data so it is
easier to swap.

Caching automatically triggers when red and orange event
occur. Caching on the server seems to work well, although the proxy
seems slow now. PNGs taking ~15 seconds to load. Might be updates to
VMware network though. Also, getting lots of DNS issues

Reproduction is ready! Just needs a UI. Added a new permission that
clears that browser cache before prepping the proxy cache so that we
properly catch all assets, changed the base64 formatting so that it
doesn't break for JS objects, and alsomake a need payload that is
executed inline rather than via an injected function due to the very
difficult timing issues with when a script was injected and when it
needed to be checked.

moving some columns around, removing old debugging logs, and making sure the tab closes as soon as the page is finished loading for cached requests

Signed-off-by: Jacob Ryan Heath <jake.heath@nccgroup.trust>